Create book creator strategy and change the big switch case part
