+++ src/main/java/com/reka/lakatos/searchbyisbn/crawler/ervinszabolibrary/bookcreation/strategy/BookPropertyUpdatingStrategy.java
@@ -1,0 +1,8 @@
+package com.reka.lakatos.searchbyisbn.crawler.ervinszabolibrary.bookcreation.strategy;
+
+import com.reka.lakatos.searchbyisbn.document.Book;
+
+public interface BookPropertyUpdatingStrategy {
+
+    void updateProperty(Book book, String property);
+}
